+++ core/minter/minter.go
@@ -90,8 +90,9 @@ func NewMinterBlockchain() *Blockchain {
 
 	blockchain.stateCheck = state.NewForCheck(blockchain.stateDeliver)
 
-	// Set start height for rewards
+	// Set start height for rewards and validators
 	rewards.SetStartHeight(applicationDB.GetStartHeight())
+	validators.SetStartHeight(applicationDB.GetStartHeight())
 
 	return blockchain
 }
@@ -129,6 +130,7 @@ func (app *Blockchain) InitChain(req abciTypes.RequestInitChain) abciTypes.Respo
 	app.appDB.SetStartHeight(genesisState.StartHeight)
 	app.appDB.SaveValidators(vals)
 	rewards.SetStartHeight(genesisState.StartHeight)
+	validators.SetStartHeight(genesisState.StartHeight)
 
 	return abciTypes.ResponseInitChain{
 		Validators: vals,