+++ core/validators/validators.go
@@ -1,6 +1,9 @@
 package validators
 
+var startHeight uint64 = 0
+
 func GetValidatorsCountForBlock(block uint64) int {
+	block += startHeight
 	count := 16 + (block/518400)*4
 
 	if count > 256 {
@@ -12,4 +15,8 @@ func GetValidatorsCountForBlock(block uint64) int {
 
 func GetCandidatesCountForBlock(block uint64) int {
 	return GetValidatorsCountForBlock(block) * 3
+}
+
+func SetStartHeight(sHeight uint64) {
+	startHeight = sHeight
 }